reorganized files, added mission planner client
