--- client/interop.txt
@@ -1,22 +1,0 @@
-Django==1.9
-Pillow==3.1.1
-argparse==1.2.1
-cycler==0.10.0
-django-debug-toolbar==1.4
-enum==0.4.6
-enum34==1.1.2
-gevent==1.0.2
-greenlet==0.4.9
-iso8601==0.1.11
-matplotlib==1.5.1
-numpy==1.10.4
-psycopg2==2.6.1
-pyparsing==2.1.0
-pysendfile==2.0.1
-python-dateutil==2.4.2
-pytz==2015.7
-scipy==0.17.0
-simplekml==1.2.8
-six==1.10.0
-sqlparse==0.1.18
-wsgiref==0.1.2
